Move generic run utils in top level dir to avoid import errors
