+++ test/upgrade/continual/broker.go
@@ -36,7 +36,7 @@ import (
 	eventingkafkaupgrade "knative.dev/eventing-kafka/test/upgrade/continual"
 
 	"knative.dev/eventing-kafka-broker/control-plane/pkg/kafka"
-	testingpkg "knative.dev/eventing-kafka-broker/test/pkg/testing"
+	testingpkg "knative.dev/eventing-kafka-broker/test/pkg"
 )
 
 const (
